Change location of stack_batch
