+++ src/otx/algo/detection/atss.py
@@ -18,6 +18,7 @@ from otx.algo.detection.heads.atss_head import ATSSHead
 from otx.algo.detection.necks.fpn import FPN
 from otx.algo.detection.ssd import SingleStageDetector
 from otx.algo.utils.mmconfig import read_mmconfig
+from otx.algo.utils.mmengine_utils import stack_batch
 from otx.algo.utils.support_otx_v1 import OTXv1Helper
 from otx.core.config.data import TileConfig
 from otx.core.data.entity.base import OTXBatchLossEntity
@@ -103,6 +104,8 @@ class ATSS(ExplainableOTXDetModel):
         return model
 
     def _customize_inputs(self, entity: DetBatchDataEntity) -> dict[str, Any]:
+        if isinstance(entity.images, list):
+            entity.images = stack_batch(entity.images, pad_size_divisor=32)
         inputs: dict[str, Any] = {}
 
         inputs["entity"] = entity
