+++ src/otx/algo/detection/yolox.py
@@ -17,6 +17,7 @@ from otx.algo.detection.heads.yolox_head import YOLOXHead
 from otx.algo.detection.necks.yolox_pafpn import YOLOXPAFPN
 from otx.algo.detection.ssd import SingleStageDetector
 from otx.algo.utils.mmconfig import read_mmconfig
+from otx.algo.utils.mmengine_utils import stack_batch
 from otx.algo.utils.support_otx_v1 import OTXv1Helper
 from otx.core.config.data import TileConfig
 from otx.core.data.entity.base import OTXBatchLossEntity
@@ -115,6 +116,8 @@ class OTXYOLOX(ExplainableOTXDetModel):
         return detector
 
     def _customize_inputs(self, entity: DetBatchDataEntity) -> dict[str, Any]:
+        if isinstance(entity.images, list):
+            entity.images = stack_batch(entity.images, pad_size_divisor=32, pad_value=114)
         inputs: dict[str, Any] = {}
 
         inputs["entity"] = entity
