+++ src/otx/core/data/entity/base.py
@@ -6,6 +6,7 @@
 
 from __future__ import annotations
 
+import warnings
 from collections.abc import Mapping
 from dataclasses import asdict, dataclass, field, fields
 from typing import TYPE_CHECKING, Any, Dict, Generic, Iterator, TypeVar
@@ -607,9 +608,19 @@ class OTXBatchDataEntity(Generic[T_OTXDataEntity]):
         images = [entity.image for entity in entities]
         like = next(iter(images))
 
+        if stack_images and not all(like.shape == entity.image.shape for entity in entities):  # type: ignore[union-attr]
+            msg = (
+                "You set stack_images as True, but not all images in the batch has same shape. "
+                "In this case, we cannot stack images. Some tasks, e.g., detection, "
+                "can have different image shapes among samples in the batch. However, if it is not your intention, "
+                "consider setting stack_images as False in the config."
+            )
+            warnings.warn(msg, stacklevel=1)
+            stack_images = False
+
         return OTXBatchDataEntity(
             batch_size=batch_size,
-            images=tv_tensors.wrap(cls.stack_batch(images), like=like) if stack_images else images,
+            images=tv_tensors.wrap(torch.stack(images), like=like) if stack_images else images,
             imgs_info=[entity.img_info for entity in entities],
         )
 
@@ -626,52 +637,6 @@ class OTXBatchDataEntity(Generic[T_OTXDataEntity]):
 
         like = next(iter(self.images))
         return tv_tensors.wrap(stack(self.images), like=like)
-
-    @staticmethod
-    # TODO(someone): Pad size divisior and pad value should be configurable
-    def stack_batch(
-        tensor_list: list[torch.Tensor],
-        pad_size_divisor: int = 1,
-        pad_value: int | float = 0,
-    ) -> torch.Tensor:
-        """Stack multiple tensors to form a batch.
-
-        Pad the tensor to the max shape use the right bottom padding mode in these images.
-        If ``pad_size_divisor > 0``, add padding to ensure the shape of each dim is
-        divisible by ``pad_size_divisor``.
-
-        Args:
-            tensor_list (List[Tensor]): A list of tensors with the same dim.
-            pad_size_divisor (int): If ``pad_size_divisor > 0``, add padding
-                to ensure the shape of each dim is divisible by
-                ``pad_size_divisor``. This depends on the model, and many
-                models need to be divisible by 32. Defaults to 1
-            pad_value (int, float): The padding value. Defaults to 0.
-
-        Returns:
-            Tensor: The n dim tensor.
-        """
-        dim = tensor_list[0].dim()
-        num_img = len(tensor_list)
-        all_sizes: torch.Tensor = torch.Tensor([tensor.shape for tensor in tensor_list])
-        max_sizes = torch.ceil(torch.max(all_sizes, dim=0)[0] / pad_size_divisor) * pad_size_divisor
-        padded_sizes = max_sizes - all_sizes
-        # The first dim normally means channel,  which should not be padded.
-        padded_sizes[:, 0] = 0
-        if padded_sizes.sum() == 0:
-            return torch.stack(tensor_list)
-        # `pad` is the second arguments of `F.pad`. If pad is (1, 2, 3, 4),
-        # it means that padding the last dim with 1(left) 2(right), padding the
-        # penultimate dim to 3(top) 4(bottom). The order of `pad` is opposite of
-        # the `padded_sizes`. Therefore, the `padded_sizes` needs to be reversed,
-        # and only odd index of pad should be assigned to keep padding "right" and
-        # "bottom".
-        pad = torch.zeros(num_img, 2 * dim, dtype=torch.int)
-        pad[:, 1::2] = padded_sizes[:, range(dim - 1, -1, -1)]
-        batch_tensor = []
-        for idx, tensor in enumerate(tensor_list):
-            batch_tensor.append(torch.nn.functional.pad(tensor, tuple(pad[idx].tolist()), value=pad_value))
-        return torch.stack(batch_tensor)
 
     def pin_memory(self: T_OTXBatchDataEntity) -> T_OTXBatchDataEntity:
         """Pin memory for member tensor variables."""
